Move utility method ResourceTest#setBuildOrder() to only consumers #903
The method ResourceTest#setBuildOrder() has two consuming paths in the
inheritance hierarchy, one concrete test class and one abstract test
class that is lower in the type hierarchy. This change pushes down the
utility method to clean up the REsourceTest inheritance hierarchy.

Contributes to
https://github.com/eclipse-platform/eclipse.platform/issues/903
